update SignupController (post)

+++ starter/cloudstorage/src/main/java/com/udacity/jwdnd/course1/cloudstorage/model/User.java
@@ -2,14 +2,14 @@ package com.udacity.jwdnd.course1.cloudstorage.model;
 
 public class User {
 
-  private int id;
+  private Long id;
   private String username;
   private String firstName;
   private String lastName;
   private String salt;
   private String password;
 
-  public User(int id, String username, String firstName, String lastName, String salt, String password) {
+  public User(Long id, String username, String firstName, String lastName, String salt, String password) {
     this.id = id;
     this.username = username;
     this.firstName = firstName;
@@ -18,11 +18,11 @@ public class User {
     this.password = password;
   }
 
-  public int getId() {
+  public Long getId() {
     return id;
   }
 
-  public void setId(int id) {
+  public void setId(Long id) {
     this.id = id;
   }
